+++ MapboxCoreNavigation/RouteProgress.swift
@@ -196,7 +196,7 @@ open class RouteLegProgress: NSObject {
     /**
      An approximatation of the closest segment on the current leg.
      */
-    @objc public var currentSegment: Int {
+    @objc public var currentSegmentIndex: Int {
         let legCoordinates = Array(leg.steps.flatMap { $0.coordinates }.joined())
         return Int(Double(legCoordinates.count) * fractionTraveled)
     }
